// src/app/shared/models/movie.ts
export class Movie{
    movieId?: number;
    movieName?: string;
	movieActor?: number;
    movieDuration?: number;
    movieGerne?: number;
    movieOverview?:String;
    movieUrl?:String;
    movieReview?:number;
    movieCost?:number;
    movieReleaseYear?:number;
}
    

	
	
	
	
	
	
	
	
	

	
	
	
	
	